resolved case where customer was undefined
